Improved chunk loading
The chunk loading has been somewhat improved and optimised. The face of
a block in a chunk is now only rendered if there is no neighbour in the
direction of that face. This drastically reduces the number of triangles
drawn.

I also started implementing a way to convert from world coordinates to
chunk indexes. This will be needed to implement a world that is
virtually infinitely generating.

TODO: Offset coordinate system by the player position when computing
chunk indexes. Only include chunks within a certain distance from the
player.

TODO: add landscape generation via noise algorithms.