Simple lighting and refactored a ton of stuff
Refactored the camera and controls.
Refactored the chunk structure from a 3d structure in a 1d array to a
fxied-size 2d structure in a 2d array.

I also wrote a simple Phong shader and found the issue with the atefacts
along the vertex lines. It was an iterpolation error when passing the
varying vectors from the vertex to the fragment shader; it did not stay
normalised. I moved the normalisation to the fragment shader and the
artefacts disappeared.