Improved chunk loading once again
I improved the chunk loading with a sort of "occlusion chunk loading"
algorith; chunks are only setup (a very expensve process) if their
neighbour chunks in the direction of the player are already setup AND
have no full face towards the chunk that is being considered for setup.
This prevents chunks from loading beneath the top layer. I also added a
radius below which chunks are forced to be setup, to make sure chunks
immediately arround the player are always loaded. The result is a
massive reduction in the voxels that are setup, and hence in the memory
used. This allows a greater view distance since more chunks can now be
loaded - combined with a larger chunk size (32*32*32) the view distance
is actually kind of decent now.

I also added some more sexy stats to the gui.