Refactored voxel generation and ... more chunk loading!
I refactored the chunk loading to a more modular design.

I also improved the chunk loading again. For real this time though. The
neighbour of a chunk is now only loaded if the face towards that chunk
is not completely full of voxels (if it is, then there's no reason to
load cause it's not going to be visible anyway). This saves *SO* much
memory, and it is by far the simplest (good) strategy I have been able
to come up with.